requirements.txt
google-generativeai
faiss-cpu
python-dotenv
numpy
pandas